Changed names of classes and variables
Also changed the way lists appear on the site, so now it will render using
javascript but not html markup. Names of variables and classes were
changed due to due to difficulties in distinguishing the names of list
elements from the names of classes and types of mods as as long as I
remember.